Feat: calculadora com float

+++ 02-pratica/calculadora_modularizada/operacoes.py
@@ -1,23 +1,23 @@
 def soma(num1,num2):
     num1 = input('Digite um número: ')
     num2 = input('Digite outro número: ')
-    resultado = int(num1) + int(num2)
+    resultado = float(num1) + float(num2)
     return resultado
 
 def subtracao(num1,num2):
     num1 = input('Digite um número: ')
     num2 = input('Digite outro número: ')
-    resultado = int(num1) - int(num2)
+    resultado = float(num1) - float(num2)
     return resultado
 
 def multiplicacao(num1,num2):
     num1 = input('Digite um número: ')
     num2 = input('Digite outro número: ')
-    resultado = int(num1) * int(num2)
+    resultado = float(num1) * float(num2)
     return resultado
 
 def divisao(num1,num2):
     num1 = input('Digite um número: ')
     num2 = input('Digite outro número: ')
-    resultado = int(num1) / int(num2)
+    resultado = float(num1) / float(num2)
     return resultado
